Get all available gene panels in a database

Get all available gene panels in a database

all_available_panels()

Value

a dataframe of metadata regarding each available panel

Examples

get_cbioportal_db("public")
all_available_panels()
#> Warning: No CBIOPORTAL_TOKEN in `.Renviron`. Try `usethis::edit_r_environ()` to add a token
#> # A tibble: 53 × 2
#>    description                                                       genePanelId
#>    <chr>                                                             <chr>      
#>  1 Targeted (27 cancer genes) sequencing of adenoid cystic carcinom… ACYC_FMI_27
#>  2 Targeted panel of 8 genes.                                        AmpliSeq   
#>  3 ARCHER-SOLID Gene Panel (62 genes)                                ARCHER-SOL…
#>  4 Targeted sequencing of various tumor types via bait v3.           bait_v3    
#>  5 Targeted sequencing of various tumor types via bait v4.           bait_v4    
#>  6 Targeted sequencing of various tumor types via bait v5.           bait_v5    
#>  7 Foundation Medicine T5a gene panel (323 genes)                    FMI-T5a    
#>  8 Foundation Medicine T7 gene panel (429 genes)                     FMI-T7     
#>  9 Foundation Medicine T5 gene panel (326 genes)                     glioma_msk…
#> 10 Foundation Medicine T7 gene panel (434 genes)                     glioma_msk…
#> # … with 43 more rows