Get Available Gene Panels For a Database

Get Available Gene Panels For a Database

available_gene_panels(base_url = NULL)

Arguments

base_url

The database URL to query. If NULL will default to URL set with set_cbioportal_db(<your_db>)

Value

a dataframe of metadata regarding each available panel

Examples

set_cbioportal_db("public")
#> ✔ You are successfully connected!
#> ✔ base_url for this R session is now set to "www.cbioportal.org/api" 
available_gene_panels()
#> # A tibble: 54 × 2
#>    description                                                       genePanelId
#>    <chr>                                                             <chr>      
#>  1 Targeted (27 cancer genes) sequencing of adenoid cystic carcinom… ACYC_FMI_27
#>  2 Targeted panel of 232 genes.                                      Agilent    
#>  3 Targeted panel of 8 genes.                                        AmpliSeq   
#>  4 ARCHER-HEME gene panel (199 genes)                                ARCHER-HEM…
#>  5 ARCHER-SOLID Gene Panel (62 genes)                                ARCHER-SOL…
#>  6 Targeted sequencing of various tumor types via bait v3.           bait_v3    
#>  7 Targeted sequencing of various tumor types via bait v5.           bait_v5    
#>  8 Targeted panel of 387 cancer-related genes.                       bcc_unige_…
#>  9 Research (CMO) IMPACT-Heme gene panel version 3.                  HemePACT_v3
#> 10 Targeted sequencing of 503 cancer-associated genes on Illumina H… DFCI_504   
#> # … with 44 more rows