Get Available Genomic Profiles For a Study or Database

Get Available Genomic Profiles For a Study or Database

available_profiles(study_id = NULL, base_url = NULL)

Arguments

study_id

A character vector of length 1 indicating study_id. See get_studies() to see all available studies for your URL. If NULL, it will return all profiles for your current database url

base_url

The database URL to query. If NULL will default to URL set with set_cbioportal_db(<your_db>)

Value

A dataframe of available genetic profiles and their profile ids

Examples

set_cbioportal_db("public")
#> ✔ You are successfully connected!
#> ✔ base_url for this R session is now set to "www.cbioportal.org/api" 
available_profiles()
#> ℹ No `study_id` provided. Returning all available genomic profiles for <www.cbioportal.org/api>
#> # A tibble: 1,657 × 11
#>    molecularAlteration… datatype name  description showProfileInAn… patientLevel
#>    <chr>                <chr>    <chr> <chr>       <lgl>            <lgl>       
#>  1 PROTEIN_LEVEL        LOG2-VA… Prot… Protein ex… FALSE            FALSE       
#>  2 PROTEIN_LEVEL        Z-SCORE  Prot… Protein ex… TRUE             FALSE       
#>  3 COPY_NUMBER_ALTERAT… DISCRETE Puta… Putative c… TRUE             FALSE       
#>  4 COPY_NUMBER_ALTERAT… CONTINU… Capp… Capped rel… FALSE            FALSE       
#>  5 MUTATION_EXTENDED    MAF      Muta… Mutation d… TRUE             FALSE       
#>  6 METHYLATION          CONTINU… Meth… Methylatio… FALSE            FALSE       
#>  7 MRNA_EXPRESSION      CONTINU… mRNA… mRNA gene … FALSE            FALSE       
#>  8 MRNA_EXPRESSION      Z-SCORE  mRNA… mRNA expre… TRUE             FALSE       
#>  9 MRNA_EXPRESSION      Z-SCORE  mRNA… Log-transf… TRUE             FALSE       
#> 10 COPY_NUMBER_ALTERAT… DISCRETE Copy… Putative c… TRUE             FALSE       
#> # … with 1,647 more rows, and 5 more variables: molecularProfileId <chr>,
#> #   studyId <chr>, genericAssayType <chr>, pivotThreshold <dbl>,
#> #   sortOrder <chr>
available_profiles(study_id = "acc_tcga")
#> # A tibble: 9 × 8
#>   molecularAlterationT… datatype name  description showProfileInAn… patientLevel
#>   <chr>                 <chr>    <chr> <chr>       <lgl>            <lgl>       
#> 1 PROTEIN_LEVEL         LOG2-VA… Prot… Protein ex… FALSE            FALSE       
#> 2 PROTEIN_LEVEL         Z-SCORE  Prot… Protein ex… TRUE             FALSE       
#> 3 COPY_NUMBER_ALTERATI… DISCRETE Puta… Putative c… TRUE             FALSE       
#> 4 COPY_NUMBER_ALTERATI… CONTINU… Capp… Capped rel… FALSE            FALSE       
#> 5 MUTATION_EXTENDED     MAF      Muta… Mutation d… TRUE             FALSE       
#> 6 METHYLATION           CONTINU… Meth… Methylatio… FALSE            FALSE       
#> 7 MRNA_EXPRESSION       CONTINU… mRNA… mRNA gene … FALSE            FALSE       
#> 8 MRNA_EXPRESSION       Z-SCORE  mRNA… mRNA expre… TRUE             FALSE       
#> 9 MRNA_EXPRESSION       Z-SCORE  mRNA… Log-transf… TRUE             FALSE       
#> # … with 2 more variables: molecularProfileId <chr>, studyId <chr>