Internal Function to Get Mutations/CNA/Fusion By Study ID

Endpoints for retrieving mutation and cna data are structurally similar. This internal function allows you to pull data from either endpoint. It has logic for sensible default guesses at study_id and molecular_profile_id when those are NULL

.get_data_by_study(
  study_id = NULL,
  molecular_profile_id = NULL,
  data_type = c("mutation", "cna", "fusion"),
  base_url = NULL,
  add_hugo = TRUE
)

Arguments

study_id

A study ID to query mutations. If NULL, guesses study ID based on molecular_profile_id.

molecular_profile_id

a molecular profile to query mutations. If NULL, guesses molecular_profile_id based on study ID.

data_type

specify what type of data to return. Options aremutation, cna, fusion.

base_url

The database URL to query If NULL will default to URL set with set_cbioportal_db(<your_db>)

add_hugo

Logical indicating whether HugoSymbol should be added to your results. cBioPortal API does not return this by default (only EntrezId) but this functions default is TRUE and adds this by default.

Value

a dataframe of mutations or CNAs

Examples

# \donttest{
set_cbioportal_db("public")
#> ✔ You are successfully connected!
#> ✔ base_url for this R session is now set to "www.cbioportal.org/api" 
.get_data_by_study(study_id = "prad_msk_2019", data_type = "cna")
#> ℹ Returning all data for the "prad_msk_2019_cna" molecular profile in the "prad_msk_2019" study
#> # A tibble: 1 × 9
#>   hugoGeneSymbol entrezGeneId uniqueSampleKey  uniquePatientKey molecularProfil…
#>   <chr>                 <int> <chr>            <chr>            <chr>           
#> 1 PTEN                   5728 c19DXzM2OTI0TF9… cF9DXzM2OTI0TDp… prad_msk_2019_c…
#> # … with 4 more variables: sampleId <chr>, patientId <chr>, studyId <chr>,
#> #   alteration <int>
.get_data_by_study(study_id = "prad_msk_2019", data_type = "mutation")
#> ℹ Returning all data for the "prad_msk_2019_mutations" molecular profile in the "prad_msk_2019" study
#> # A tibble: 26 × 33
#>    hugoGeneSymbol entrezGeneId uniqueSampleKey uniquePatientKey molecularProfil…
#>    <chr>                 <int> <chr>           <chr>            <chr>           
#>  1 ZFHX3                   463 c19DX004WDQyVF… cF9DX004WDQyVDp… prad_msk_2019_m…
#>  2 ZFHX3                   463 c19DX004WDQyVF… cF9DX004WDQyVDp… prad_msk_2019_m…
#>  3 ATR                     545 c19DX004WDQyVF… cF9DX004WDQyVDp… prad_msk_2019_m…
#>  4 BCL2                    596 c19DX004WDQyVF… cF9DX004WDQyVDp… prad_msk_2019_m…
#>  5 ETV1                   2115 c19DX1A4SzNUUl… cF9DX1A4SzNUUjp… prad_msk_2019_m…
#>  6 ETV1                   2115 c19DX1A4SzNUUl… cF9DX1A4SzNUUjp… prad_msk_2019_m…
#>  7 FAT1                   2195 c19DX004WDQyVF… cF9DX004WDQyVDp… prad_msk_2019_m…
#>  8 MSH6                   2956 c19DX1A4SzNUUl… cF9DX1A4SzNUUjp… prad_msk_2019_m…
#>  9 MSH6                   2956 c19DX1A4SzNUUl… cF9DX1A4SzNUUjp… prad_msk_2019_m…
#> 10 FOXA1                  3169 c19DX0UwS0pGSl… cF9DX0UwS0pGSjp… prad_msk_2019_m…
#> # … with 16 more rows, and 28 more variables: sampleId <chr>, patientId <chr>,
#> #   studyId <chr>, center <chr>, mutationStatus <chr>, validationStatus <chr>,
#> #   tumorAltCount <int>, tumorRefCount <int>, normalAltCount <int>,
#> #   normalRefCount <int>, startPosition <int>, endPosition <int>,
#> #   referenceAllele <chr>, proteinChange <chr>, mutationType <chr>,
#> #   functionalImpactScore <chr>, fisValue <dbl>, linkXvar <chr>, linkPdb <chr>,
#> #   linkMsa <chr>, ncbiBuild <chr>, variantType <chr>, keyword <chr>, …
.get_data_by_study(study_id = "prad_msk_2019", data_type = "fusion")
#> ℹ Returning all data for the "prad_msk_2019_fusion" molecular profile in the "prad_msk_2019" study
#> # A tibble: 4 × 17
#>   uniqueSampleKey   uniquePatientKey molecularProfil… sampleId patientId studyId
#>   <chr>             <chr>            <chr>            <chr>    <chr>     <chr>  
#> 1 c19DX0NBVVdUN19Q… cF9DX0NBVVdUNzp… prad_msk_2019_f… s_C_CAU… p_C_CAUW… prad_m…
#> 2 c19DX0RVNkVDQ19Q… cF9DX0RVNkVDQzp… prad_msk_2019_f… s_C_DU6… p_C_DU6E… prad_m…
#> 3 c19DX1ZDNlA5QV9Q… cF9DX1ZDNlA5QTp… prad_msk_2019_f… s_C_VC6… p_C_VC6P… prad_m…
#> 4 c19DX1ZDNlA5QV9Q… cF9DX1ZDNlA5QTp… prad_msk_2019_f… s_C_VC6… p_C_VC6P… prad_m…
#> # … with 11 more variables: site1EntrezGeneId <int>, site1HugoSymbol <chr>,
#> #   site1Chromosome <chr>, site1Position <int>, site2EntrezGeneId <int>,
#> #   site2HugoSymbol <chr>, ncbiBuild <chr>, center <chr>, eventInfo <chr>,
#> #   variantClass <chr>, comments <chr>

.get_data_by_study(molecular_profile_id = "prad_msk_2019_cna", data_type = "cna")
#> ℹ Returning all data for the "prad_msk_2019_cna" molecular profile in the "prad_msk_2019" study
#> # A tibble: 1 × 9
#>   hugoGeneSymbol entrezGeneId uniqueSampleKey  uniquePatientKey molecularProfil…
#>   <chr>                 <int> <chr>            <chr>            <chr>           
#> 1 PTEN                   5728 c19DXzM2OTI0TF9… cF9DXzM2OTI0TDp… prad_msk_2019_c…
#> # … with 4 more variables: sampleId <chr>, patientId <chr>, studyId <chr>,
#> #   alteration <int>
.get_data_by_study(molecular_profile_id = "prad_msk_2019_mutations", data_type = "mutation")
#> ℹ Returning all data for the "prad_msk_2019_mutations" molecular profile in the "prad_msk_2019" study
#> # A tibble: 26 × 33
#>    hugoGeneSymbol entrezGeneId uniqueSampleKey uniquePatientKey molecularProfil…
#>    <chr>                 <int> <chr>           <chr>            <chr>           
#>  1 ZFHX3                   463 c19DX004WDQyVF… cF9DX004WDQyVDp… prad_msk_2019_m…
#>  2 ZFHX3                   463 c19DX004WDQyVF… cF9DX004WDQyVDp… prad_msk_2019_m…
#>  3 ATR                     545 c19DX004WDQyVF… cF9DX004WDQyVDp… prad_msk_2019_m…
#>  4 BCL2                    596 c19DX004WDQyVF… cF9DX004WDQyVDp… prad_msk_2019_m…
#>  5 ETV1                   2115 c19DX1A4SzNUUl… cF9DX1A4SzNUUjp… prad_msk_2019_m…
#>  6 ETV1                   2115 c19DX1A4SzNUUl… cF9DX1A4SzNUUjp… prad_msk_2019_m…
#>  7 FAT1                   2195 c19DX004WDQyVF… cF9DX004WDQyVDp… prad_msk_2019_m…
#>  8 MSH6                   2956 c19DX1A4SzNUUl… cF9DX1A4SzNUUjp… prad_msk_2019_m…
#>  9 MSH6                   2956 c19DX1A4SzNUUl… cF9DX1A4SzNUUjp… prad_msk_2019_m…
#> 10 FOXA1                  3169 c19DX0UwS0pGSl… cF9DX0UwS0pGSjp… prad_msk_2019_m…
#> # … with 16 more rows, and 28 more variables: sampleId <chr>, patientId <chr>,
#> #   studyId <chr>, center <chr>, mutationStatus <chr>, validationStatus <chr>,
#> #   tumorAltCount <int>, tumorRefCount <int>, normalAltCount <int>,
#> #   normalRefCount <int>, startPosition <int>, endPosition <int>,
#> #   referenceAllele <chr>, proteinChange <chr>, mutationType <chr>,
#> #   functionalImpactScore <chr>, fisValue <dbl>, linkXvar <chr>, linkPdb <chr>,
#> #   linkMsa <chr>, ncbiBuild <chr>, variantType <chr>, keyword <chr>, …
.get_data_by_study(molecular_profile_id = "prad_msk_2019_fusion", data_type = "fusion")
#> ℹ Returning all data for the "prad_msk_2019_fusion" molecular profile in the "prad_msk_2019" study
#> # A tibble: 4 × 17
#>   uniqueSampleKey   uniquePatientKey molecularProfil… sampleId patientId studyId
#>   <chr>             <chr>            <chr>            <chr>    <chr>     <chr>  
#> 1 c19DX0NBVVdUN19Q… cF9DX0NBVVdUNzp… prad_msk_2019_f… s_C_CAU… p_C_CAUW… prad_m…
#> 2 c19DX0RVNkVDQ19Q… cF9DX0RVNkVDQzp… prad_msk_2019_f… s_C_DU6… p_C_DU6E… prad_m…
#> 3 c19DX1ZDNlA5QV9Q… cF9DX1ZDNlA5QTp… prad_msk_2019_f… s_C_VC6… p_C_VC6P… prad_m…
#> 4 c19DX1ZDNlA5QV9Q… cF9DX1ZDNlA5QTp… prad_msk_2019_f… s_C_VC6… p_C_VC6P… prad_m…
#> # … with 11 more variables: site1EntrezGeneId <int>, site1HugoSymbol <chr>,
#> #   site1Chromosome <chr>, site1Position <int>, site2EntrezGeneId <int>,
#> #   site2HugoSymbol <chr>, ncbiBuild <chr>, center <chr>, eventInfo <chr>,
#> #   variantClass <chr>, comments <chr>
# }