Get Gene Name Alias for a Given Hugo Symbol

This function grabs known gene aliases for a given Hugo Symbol. You may notice that genes -alias pairs are not always consistent. For example get_alias("KMT2D") will return "MLL2" but get_alias("MLL2") will not return "KMT2D" This function relies on the existing cBioPortal API which controls this database of aliases. Therefore, this is a convenience function but you may want to consider a more carefully curated alias list like cbioportalR::impact_gene_info

Usage

get_alias(hugo_symbol = NULL, base_url = NULL)

Arguments

hugo_symbol

a hugo symbol for which to return aliases

base_url

The database URL to query

Value

A character string with all aliases

Examples

# \dontrun{

get_alias(hugo_symbol = "FGFR3", base_url = 'www.cbioportal.org/api')
#> # A tibble: 4 × 2
#>   hugo_symbol alias
#>   <chr>       <chr>
#> 1 FGFR3       ACH  
#> 2 FGFR3       CD333
#> 3 FGFR3       CEK2 
#> 4 FGFR3       JTK4 
get_alias(hugo_symbol = c("FGFR3", "TP53"), base_url = 'www.cbioportal.org/api')
#> # A tibble: 6 × 2
#>   hugo_symbol alias
#>   <chr>       <chr>
#> 1 FGFR3       ACH  
#> 2 FGFR3       CD333
#> 3 FGFR3       CEK2 
#> 4 FGFR3       JTK4 
#> 5 TP53        LFS1 
#> 6 TP53        p53  
# }