Get A List of Genes for a Specified Database

Get A List of Genes for a Specified Database

Usage

get_genes(base_url = NULL)

Arguments

base_url

The database URL to query If NULL will default to URL set with set_cbioportal_db(<your_db>)

Value

A dataframe of gene ids, hugo symbols, and gene types

Examples

# \dontrun{
get_genes(base_url = 'www.cbioportal.org/api')
#> # A tibble: 46,167 × 3
#>    entrezGeneId hugoGeneSymbol type          
#>           <int> <chr>          <chr>         
#>  1    101340252 SNORD121B      snoRNA        
#>  2         3360 HTR4           protein-coding
#>  3    100113379 SNORD121A      snoRNA        
#>  4        -3510 MIR-124A-1/5P  miRNA         
#>  5         3362 HTR6           protein-coding
#>  6         3363 HTR7           protein-coding
#>  7    106481909 RNU6-654P      pseudogene    
#>  8        56987 BBX            protein-coding
#>  9    106479427 RN7SL563P      pseudogene    
#> 10    107985793 ARL14EPP1      pseudogene    
#> # ℹ 46,157 more rows
# }