Get gene mutations by sample ID or Study

Get gene mutations by sample ID or Study

get_mutations(sample_ids = NULL, study_id = NULL, panel = NULL, genes = NULL)

Arguments

sample_ids
A character vector of sample ids

study_id
A character vector of length 1 indicating study_id. See get_studies() to see available studies.

panel
OPTIONAL argument. A character vector of length 1 indicating a specific panel to be used. If not NULL, the panel will be looked up with get_panel() and only genes in that panel will be returned.

genes
A list of genes to query. default is all impact genes.

Value

A dataframe of mutations for each sample ID (in maf file format)

Examples

get_cbioportal_db("public")
get_mutations(sample_ids = c("TCGA-19-5959-01", "TCGA-18-4083-01"))

#> No genes specified by user, so querying curated list of impact genes.
#>               To view queried gene list run `cbioportalR::impact_gene_info`

#> # A tibble: 16 x 39
#>    uniqueSampleKey uniquePatientKey molecularProfil… sampleId patientId
#>    <chr>           <chr>            <chr>            <chr>    <chr>    
#>  1 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#>  2 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#>  3 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#>  4 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#>  5 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#>  6 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#>  7 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#>  8 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#>  9 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#> 10 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#> 11 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#> 12 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#> 13 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#> 14 VENHQS0xOC00MD… VENHQS0xOC00MDg… lusc_tcga_pan_c… TCGA-18… TCGA-18-…
#> 15 VENHQS0xOS01OT… VENHQS0xOS01OTU… gbm_tcga_pan_ca… TCGA-19… TCGA-19-…
#> 16 VENHQS0xOS01OT… VENHQS0xOS01OTU… gbm_tcga_pan_ca… TCGA-19… TCGA-19-…
#> # … with 34 more variables: entrezGeneId <int>, studyId <chr>,
#> #   driverFilter <chr>, driverFilterAnnotation <chr>, driverTiersFilter <chr>,
#> #   driverTiersFilterAnnotation <chr>, center <chr>, mutationStatus <chr>,
#> #   validationStatus <chr>, tumorAltCount <int>, tumorRefCount <int>,
#> #   normalAltCount <int>, normalRefCount <int>, startPosition <int>,
#> #   endPosition <int>, referenceAllele <chr>, proteinChange <chr>,
#> #   mutationType <chr>, functionalImpactScore <chr>, fisValue <dbl>,
#> #   linkXvar <chr>, linkPdb <chr>, linkMsa <chr>, ncbiBuild <chr>,
#> #   variantType <chr>, keyword <chr>, chr <chr>, variantAllele <chr>,
#> #   refseqMrnaId <chr>, proteinPosStart <int>, proteinPosEnd <int>,
#> #   HGVSp_Short <chr>, Protein_position <chr>, Hugo_Symbol <chr>