Get Mutations By Sample ID

Get Mutations By Sample ID

Usage

get_mutations_by_sample(
  sample_id = NULL,
  study_id = NULL,
  molecular_profile_id = NULL,
  sample_study_pairs = NULL,
  genes = NULL,
  panel = NULL,
  add_hugo = NULL,
  base_url = NULL
)

Arguments

sample_id

a vector of sample IDs (character)

study_id

A string indicating the study ID from which to pull data. If no study ID, will guess the study ID based on your URL and inform. Only 1 study ID can be passed. If mutations/cna from more than 1 study needed, see sample_study_pairs

molecular_profile_id

A string indicating the molecular profile ID from which to pull data. If ID supplied, will guess the molecular profile ID based on the study ID. Only 1 molecular profile ID can be passed. If mutations from more than 1 study needed, see sample_study_pairs

sample_study_pairs

A dataframe with columns: sample_id, study_id and molecular_profile_id (optional). Variations in capitalization of column names are accepted. This can be used in place of sample_id, study_id, molecular_profile_id arguments above if you need to pull samples from several different studies at once. If passed this will take overwrite sample_id, study_id, molecular_profile_id if also passed.

genes

A vector of Entrez ids or Hugo symbols. If Hugo symbols are supplied, they will be converted to entrez ids using the get_entrez_id() function. If panel and genes are both NULL (default), it will return gene results for all available genomic data for that sample.

panel

One or more panel IDs to query (e.g. 'IMPACT468'). If panel and genes are both NULL (default), it will return gene results for all available genomic data for that sample.

add_hugo

Logical indicating whether HugoSymbol should be added to your results. cBioPortal API does not return this by default (only EntrezId) but this function's default is TRUE and adds this by default.

base_url

The database URL to query If NULL will default to URL set with set_cbioportal_db(<your_db>)

Value

A data frame of mutations (maf file format)

Examples

get_mutations_by_sample(sample_id = c("TCGA-OR-A5J2-01","TCGA-OR-A5J6-01"),
study_id = "acc_tcga",
base_url = "public")
#> The following parameters were used in query:
#> Study ID: "acc_tcga"
#> Molecular Profile ID: "acc_tcga_mutations"
#> Genes: "All available genes"
#> # A tibble: 173 × 33
#>    hugoGeneSymbol entrezGeneId uniqueSampleKey uniquePatientKey molecularProfil…
#>    <chr>                 <int> <chr>           <chr>            <chr>           
#>  1 ZFPM1                161882 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  2 ZFPM1                161882 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  3 ZFPM1                161882 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  4 ZFPM1                161882 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  5 ZNF787               126208 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  6 PODXL                  5420 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  7 CCDC102A              92922 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  8 TVP23C               201158 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  9 RINL                 126432 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#> 10 ZNF628                89887 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#> # … with 163 more rows, and 28 more variables: sampleId <chr>, patientId <chr>,
#> #   studyId <chr>, center <chr>, mutationStatus <chr>, validationStatus <chr>,
#> #   tumorAltCount <int>, tumorRefCount <int>, normalAltCount <int>,
#> #   normalRefCount <int>, startPosition <int>, endPosition <int>,
#> #   referenceAllele <chr>, proteinChange <chr>, mutationType <chr>,
#> #   functionalImpactScore <chr>, fisValue <dbl>, linkXvar <chr>, linkPdb <chr>,
#> #   linkMsa <chr>, ncbiBuild <chr>, variantType <chr>, keyword <chr>, …