Get Copy Number Segmentation Data By Study

Get Copy Number Segmentation Data By Study

Usage

get_segments_by_study(study_id = NULL, add_hugo = TRUE, base_url = NULL)

Arguments

study_id

A study ID to query mutations. If NULL, guesses study ID based on molecular_profile_id.

add_hugo

Logical indicating whether HugoGeneSymbol should be added to your resulting data frame, if not already present in raw API results. Argument is TRUE by default. If FALSE, results will be returned as is (i.e. any existing Hugo Symbol columns in raw results will not be removed).

base_url

The database URL to query If NULL will default to URL set with set_cbioportal_db(<your_db>)

Value

A dataframe of CNA segments

Examples

# \dontrun{
get_segments_by_study(study_id = "prad_msk_2019")
#> ℹ Returning all "copy number segmentation" data for the "prad_msk_2019" study
#> # A tibble: 714 × 10
#>    uniqueSampleKey  uniquePatientKey patientId  start    end segmentMean studyId
#>    <chr>            <chr>            <chr>      <int>  <int>       <dbl> <chr>  
#>  1 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… p_C_03LN… 2.49e6 2.47e8      0.0061 prad_m…
#>  2 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… p_C_03LN… 4.72e6 1.35e8     -0.0516 prad_m…
#>  3 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… p_C_03LN… 1.37e8 1.37e8      0.410  prad_m…
#>  4 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… p_C_03LN… 1.42e8 2.43e8      0.054  prad_m…
#>  5 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… p_C_03LN… 1.45e6 1.94e8     -0.0092 prad_m…
#>  6 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… p_C_03LN… 1.40e6 1.90e8     -0.0331 prad_m…
#>  7 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… p_C_03LN… 2.19e5 1.80e8      0.0139 prad_m…
#>  8 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… p_C_03LN… 3.93e5 3.13e7     -0.013  prad_m…
#>  9 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… p_C_03LN… 3.13e7 3.13e7      0.958  prad_m…
#> 10 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… p_C_03LN… 3.13e7 1.71e8      0.0438 prad_m…
#> # ℹ 704 more rows
#> # ℹ 3 more variables: sampleId <chr>, chromosome <chr>, numberOfProbes <int>
get_segments_by_study(molecular_profile_id = "prad_msk_2019_cna")
#> Error in get_segments_by_study(molecular_profile_id = "prad_msk_2019_cna"): unused argument (molecular_profile_id = "prad_msk_2019_cna")
# }