Get Metadata on All Available Studies in Database or a Specified Study

Get Metadata on All Available Studies in Database or a Specified Study

Usage

get_study_info(study_id = NULL, base_url = NULL)

Arguments

study_id

one or more study IDs (see available_studies() to lookup IDs)

base_url

The database URL to query. If NULL will default to URL set with set_cbioportal_db(<your_db>)

Value

A dataframe of study metadata

Examples

# \dontrun{
set_cbioportal_db("public")
#> ✔ You are successfully connected!
#> ✔ base_url for this R session is now set to "www.cbioportal.org/api" 
get_study_info("acc_tcga")
#>                                               name
#> 1 Adrenocortical Carcinoma (TCGA, Firehose Legacy)
#>                                                                                                                                                                                     description
#> 1 TCGA Adrenocortical Carcinoma. Source data from <A HREF="http://gdac.broadinstitute.org/runs/stddata__2016_01_28/data/ACC/20160128/">GDAC Firehose</A>. Previously known as TCGA Provisional.
#>   publicStudy groups status          importDate allSampleCount
#> 1        TRUE PUBLIC      0 2024-01-19 09:11:57             92
#>   sequencedSampleCount cnaSampleCount mrnaRnaSeqSampleCount
#> 1                   90             90                     0
#>   mrnaRnaSeqV2SampleCount mrnaMicroarraySampleCount miRnaSampleCount
#> 1                      79                         0                0
#>   methylationHm27SampleCount rppaSampleCount massSpectrometrySampleCount
#> 1                          0              46                           0
#>   completeSampleCount readPermission treatmentCount structuralVariantCount
#> 1                  75           TRUE              0                      0
#>    studyId cancerTypeId          cancerType.name cancerType.dedicatedColor
#> 1 acc_tcga          acc Adrenocortical Carcinoma                    Purple
#>   cancerType.shortName cancerType.parent cancerType.cancerTypeId
#> 1                  ACC     adrenal_gland                     acc
#>   referenceGenome
#> 1            hg19
# }