Get Available Genomic Profiles For a Study or Database

Get Available Genomic Profiles For a Study or Database

Usage

available_profiles(study_id = NULL, base_url = NULL)

Arguments

study_id

A character vector of length 1 indicating study_id. See get_studies() to see all available studies for your URL. If NULL, it will return all profiles for your current database url

base_url

The database URL to query. If NULL will default to URL set with set_cbioportal_db(<your_db>)

Value

A dataframe of available genetic profiles and their profile ids

Examples

# \dontrun{
set_cbioportal_db("public")
#> ✔ You are successfully connected!
#> ✔ base_url for this R session is now set to "www.cbioportal.org/api" 
available_profiles()
#> ℹ No `study_id` provided. Returning all available genomic profiles for <www.cbioportal.org/api>
#> # A tibble: 2,242 × 11
#>    molecularAlterationType datatype name      description showProfileInAnalysi…¹
#>    <chr>                   <chr>    <chr>     <chr>       <lgl>                 
#>  1 COPY_NUMBER_ALTERATION  DISCRETE Putative… Putative c… TRUE                  
#>  2 MUTATION_EXTENDED       MAF      Mutations Somatic mu… TRUE                  
#>  3 STRUCTURAL_VARIANT      SV       Structur… Structural… TRUE                  
#>  4 COPY_NUMBER_ALTERATION  DISCRETE Putative… Putative c… TRUE                  
#>  5 MUTATION_EXTENDED       MAF      Mutations Comprehens… TRUE                  
#>  6 MUTATION_EXTENDED       MAF      Mutations Somatic mu… TRUE                  
#>  7 STRUCTURAL_VARIANT      SV       Structur… Structural… TRUE                  
#>  8 MUTATION_EXTENDED       MAF      Mutations WGS of 28 … TRUE                  
#>  9 STRUCTURAL_VARIANT      SV       Structur… Structural… TRUE                  
#> 10 MUTATION_EXTENDED       MAF      Mutations Mutation d… TRUE                  
#> # ℹ 2,232 more rows
#> # ℹ abbreviated name: ¹​showProfileInAnalysisTab
#> # ℹ 6 more variables: patientLevel <lgl>, molecularProfileId <chr>,
#> #   studyId <chr>, genericAssayType <chr>, pivotThreshold <dbl>,
#> #   sortOrder <chr>
available_profiles(study_id = "acc_tcga")
#> # A tibble: 9 × 8
#>   molecularAlterationType datatype   name     description showProfileInAnalysi…¹
#>   <chr>                   <chr>      <chr>    <chr>       <lgl>                 
#> 1 PROTEIN_LEVEL           LOG2-VALUE Protein… Protein ex… FALSE                 
#> 2 PROTEIN_LEVEL           Z-SCORE    Protein… Protein ex… TRUE                  
#> 3 COPY_NUMBER_ALTERATION  DISCRETE   Putativ… Putative c… TRUE                  
#> 4 COPY_NUMBER_ALTERATION  CONTINUOUS Capped … Capped rel… FALSE                 
#> 5 MUTATION_EXTENDED       MAF        Mutatio… Mutation d… TRUE                  
#> 6 METHYLATION             CONTINUOUS Methyla… Methylatio… FALSE                 
#> 7 MRNA_EXPRESSION         CONTINUOUS mRNA ex… mRNA gene … FALSE                 
#> 8 MRNA_EXPRESSION         Z-SCORE    mRNA ex… mRNA expre… TRUE                  
#> 9 MRNA_EXPRESSION         Z-SCORE    mRNA ex… Log-transf… TRUE                  
#> # ℹ abbreviated name: ¹​showProfileInAnalysisTab
#> # ℹ 3 more variables: patientLevel <lgl>, molecularProfileId <chr>,
#> #   studyId <chr>
# }