Internal Function to Get Mutations/CNA/Fusion By Sample ID

Internal Function to Get Mutations/CNA/Fusion By Sample ID

Usage

.get_data_by_sample(
  sample_id = NULL,
  study_id = NULL,
  molecular_profile_id = NULL,
  sample_study_pairs = NULL,
  data_type = c("mutation", "cna", "fusion"),
  genes = NULL,
  panel = NULL,
  add_hugo = TRUE,
  base_url = NULL
)

Arguments

sample_id

a vector of sample IDs (character)

study_id

A string indicating the study ID from which to pull data. If no study ID, will guess the study ID based on your URL and inform. Only 1 study ID can be passed. If mutations/cna from more than 1 study needed, see sample_study_pairs

molecular_profile_id

A string indicating the molecular profile ID from which to pull data. If ID supplied, will guess the molecular profile ID based on the study ID. Only 1 molecular profile ID can be passed. If mutations from more than 1 study needed, see sample_study_pairs

sample_study_pairs

A dataframe with columns: sample_id, study_id and molecular_profile_id (optional). Variations in capitalization of column names are accepted. This can be used in place of sample_id, study_id, molecular_profile_id arguments above if you need to pull samples from several different studies at once. If passed this will take overwrite sample_id, study_id, molecular_profile_id if also passed.

data_type

specify what type of data to return. Options aremutations, cna, fusion

genes

A vector of Entrez ids or Hugo symbols. If Hugo symbols are supplied, they will be converted to entrez ids using the get_entrez_id() function. If panel and genes are both NULL (default), it will return gene results for all available genomic data for that sample.

panel

One or more panel IDs to query (e.g. 'IMPACT468'). If panel and genes are both NULL (default), it will return gene results for all available genomic data for that sample.

add_hugo

Logical indicating whether HugoSymbol should be added to your results. cBioPortal API does not return this by default (only EntrezId) but this function's default is TRUE and adds this by default.

base_url

The database URL to query If NULL will default to URL set with set_cbioportal_db(<your_db>)

Value

A dataframe of mutations or CNAs

Examples

set_cbioportal_db("public")
#> ✔ You are successfully connected!
#> ✔ base_url for this R session is now set to "www.cbioportal.org/api" 
.get_data_by_sample(sample_id = c("TCGA-OR-A5J2-01","TCGA-OR-A5J6-01"),
 study_id = "acc_tcga", data_type = "mutation")
#> The following parameters were used in query:
#> Study ID: "acc_tcga"
#> Molecular Profile ID: "acc_tcga_mutations"
#> Genes: "All available genes"
#> # A tibble: 173 × 33
#>    hugoGeneSymbol entre…¹ uniqu…² uniqu…³ molec…⁴ sampl…⁵ patie…⁶ studyId center
#>    <chr>            <int> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr> 
#>  1 ZFPM1           161882 VENHQS… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… hgsc.…
#>  2 ZFPM1           161882 VENHQS… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… hgsc.…
#>  3 ZFPM1           161882 VENHQS… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… hgsc.…
#>  4 ZFPM1           161882 VENHQS… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… hgsc.…
#>  5 ZNF787          126208 VENHQS… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… hgsc.…
#>  6 PODXL             5420 VENHQS… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… hgsc.…
#>  7 CCDC102A         92922 VENHQS… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… hgsc.…
#>  8 TVP23C          201158 VENHQS… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… ucsc.…
#>  9 RINL            126432 VENHQS… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… hgsc.…
#> 10 ZNF628           89887 VENHQS… VENHQS… acc_tc… TCGA-O… TCGA-O… acc_tc… hgsc.…
#> # … with 163 more rows, 24 more variables: mutationStatus <chr>,
#> #   validationStatus <chr>, tumorAltCount <int>, tumorRefCount <int>,
#> #   normalAltCount <int>, normalRefCount <int>, startPosition <int>,
#> #   endPosition <int>, referenceAllele <chr>, proteinChange <chr>,
#> #   mutationType <chr>, functionalImpactScore <chr>, fisValue <dbl>,
#> #   linkXvar <chr>, linkPdb <chr>, linkMsa <chr>, ncbiBuild <chr>,
#> #   variantType <chr>, keyword <chr>, chr <chr>, variantAllele <chr>, …
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

.get_data_by_sample(sample_id = c("DS-sig-010-P2"),
 molecular_profile_id = "blca_plasmacytoid_mskcc_2016_cna", data_type = "cna")
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016"
#> Molecular Profile ID: "blca_plasmacytoid_mskcc_2016_cna"
#> Genes: "All available genes"
#> # A tibble: 2 × 9
#>   hugoGeneSymbol entre…¹ uniqu…² uniqu…³ molec…⁴ sampl…⁵ patie…⁶ studyId alter…⁷
#>   <chr>            <int> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>     <int>
#> 1 YAP1             10413 RFMtc2… RFMtc2… blca_p… DS-sig… DS-sig… blca_p…       2
#> 2 CD79B              974 RFMtc2… RFMtc2… blca_p… DS-sig… DS-sig… blca_p…      -2
#> # … with abbreviated variable names ¹​entrezGeneId, ²​uniqueSampleKey,
#> #   ³​uniquePatientKey, ⁴​molecularProfileId, ⁵​sampleId, ⁶​patientId, ⁷​alteration

.get_data_by_sample(sample_id = c("P-0002146-T01-IM3"),
 study_id = "blca_plasmacytoid_mskcc_2016", data_type = "fusion")
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016"
#> Molecular Profile ID: "blca_plasmacytoid_mskcc_2016_fusion"
#> Genes: "All available genes"
#> # A tibble: 1 × 15
#>   uniqueSample…¹ uniqu…² molec…³ sampl…⁴ patie…⁵ studyId site1…⁶ site1…⁷ site1…⁸
#>   <chr>          <chr>   <chr>   <chr>   <chr>   <chr>     <int> <chr>   <chr>  
#> 1 UC0wMDAyMTQ2L… UC0wMD… blca_p… P-0002… P-0002… blca_p…     999 CDH1    NA     
#> # … with 6 more variables: site1Position <int>, ncbiBuild <chr>, center <chr>,
#> #   eventInfo <chr>, variantClass <chr>, comments <chr>, and abbreviated
#> #   variable names ¹​uniqueSampleKey, ²​uniquePatientKey, ³​molecularProfileId,
#> #   ⁴​sampleId, ⁵​patientId, ⁶​site1EntrezGeneId, ⁷​site1HugoSymbol,
#> #   ⁸​site1Chromosome
#> # ℹ Use `colnames()` to see all variable names

# \donttest{
df_pairs <- data.frame(
"sample_id" = c("s_C_36924L_P001_d",
"s_C_03LNU8_P001_d"),
"study_id" = c("prad_msk_2019"))

.get_data_by_sample(sample_study_pairs = df_pairs, data_type = "mutation")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "prad_msk_2019"
#> Molecular Profile ID: prad_msk_2019_mutations
#> Genes: "All available genes"
#> # A tibble: 1 × 33
#>   hugoG…¹ entre…² uniqu…³ uniqu…⁴ molec…⁵ sampl…⁶ patie…⁷ studyId center mutat…⁸
#>   <chr>     <int> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  <chr>  
#> 1 TP53       7157 c19DXz… cF9DXz… prad_m… s_C_03… p_C_03… prad_m… MSK-I… SOMATIC
#> # … with 23 more variables: validationStatus <chr>, tumorAltCount <int>,
#> #   tumorRefCount <int>, normalAltCount <int>, normalRefCount <int>,
#> #   startPosition <int>, endPosition <int>, referenceAllele <chr>,
#> #   proteinChange <chr>, mutationType <chr>, functionalImpactScore <chr>,
#> #   fisValue <dbl>, linkXvar <chr>, linkPdb <chr>, linkMsa <chr>,
#> #   ncbiBuild <chr>, variantType <chr>, keyword <chr>, chr <chr>,
#> #   variantAllele <chr>, refseqMrnaId <chr>, proteinPosStart <int>, …
#> # ℹ Use `colnames()` to see all variable names
.get_data_by_sample(sample_study_pairs = df_pairs, genes = 7157, data_type = "mutation")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "prad_msk_2019"
#> Molecular Profile ID: prad_msk_2019_mutations
#> Genes: 7157
#> # A tibble: 1 × 33
#>   hugoG…¹ entre…² uniqu…³ uniqu…⁴ molec…⁵ sampl…⁶ patie…⁷ studyId center mutat…⁸
#>   <chr>     <int> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  <chr>  
#> 1 TP53       7157 c19DXz… cF9DXz… prad_m… s_C_03… p_C_03… prad_m… MSK-I… SOMATIC
#> # … with 23 more variables: validationStatus <chr>, tumorAltCount <int>,
#> #   tumorRefCount <int>, normalAltCount <int>, normalRefCount <int>,
#> #   startPosition <int>, endPosition <int>, referenceAllele <chr>,
#> #   proteinChange <chr>, mutationType <chr>, functionalImpactScore <chr>,
#> #   fisValue <dbl>, linkXvar <chr>, linkPdb <chr>, linkMsa <chr>,
#> #   ncbiBuild <chr>, variantType <chr>, keyword <chr>, chr <chr>,
#> #   variantAllele <chr>, refseqMrnaId <chr>, proteinPosStart <int>, …
#> # ℹ Use `colnames()` to see all variable names
.get_data_by_sample(sample_study_pairs = df_pairs, data_type = "cna")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "prad_msk_2019"
#> Molecular Profile ID: prad_msk_2019_cna
#> Genes: "All available genes"
#> # A tibble: 1 × 9
#>   hugoGeneSymbol entre…¹ uniqu…² uniqu…³ molec…⁴ sampl…⁵ patie…⁶ studyId alter…⁷
#>   <chr>            <int> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>     <int>
#> 1 PTEN              5728 c19DXz… cF9DXz… prad_m… s_C_36… p_C_36… prad_m…      -2
#> # … with abbreviated variable names ¹​entrezGeneId, ²​uniqueSampleKey,
#> #   ³​uniquePatientKey, ⁴​molecularProfileId, ⁵​sampleId, ⁶​patientId, ⁷​alteration
.get_data_by_sample(sample_study_pairs = df_pairs, data_type = "fusion")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "prad_msk_2019"
#> Molecular Profile ID: prad_msk_2019_fusion
#> Genes: "All available genes"
#> # A tibble: 0 × 0

df_pairs2 <- data.frame(
"sample_id" = c("P-0002146-T01-IM3", "s_C_CAUWT7_P001_d"),
 "study_id" = c("blca_plasmacytoid_mskcc_2016", "prad_msk_2019"))

.get_data_by_sample(sample_study_pairs = df_pairs2, data_type = "mutation")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016" and "prad_msk_2019"
#> Molecular Profile ID: blca_plasmacytoid_mskcc_2016_mutations and
#> prad_msk_2019_mutations
#> Genes: "All available genes"
#> # A tibble: 7 × 33
#>   hugoG…¹ entre…² uniqu…³ uniqu…⁴ molec…⁵ sampl…⁶ patie…⁷ studyId center mutat…⁸
#>   <chr>     <int> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  <chr>  
#> 1 TERT       7015 UC0wMD… UC0wMD… blca_p… P-0002… P-0002… blca_p… MSK-I… NA     
#> 2 TP53       7157 UC0wMD… UC0wMD… blca_p… P-0002… P-0002… blca_p… MSK-I… NA     
#> 3 NOTCH4     4855 UC0wMD… UC0wMD… blca_p… P-0002… P-0002… blca_p… MSK-I… NA     
#> 4 BLM         641 UC0wMD… UC0wMD… blca_p… P-0002… P-0002… blca_p… MSK-I… NA     
#> 5 TP53       7157 UC0wMD… UC0wMD… blca_p… P-0002… P-0002… blca_p… MSK-I… NA     
#> 6 CDKN1A     1026 UC0wMD… UC0wMD… blca_p… P-0002… P-0002… blca_p… MSK-I… NA     
#> 7 RB1        5925 UC0wMD… UC0wMD… blca_p… P-0002… P-0002… blca_p… MSK-I… NA     
#> # … with 23 more variables: validationStatus <chr>, tumorAltCount <int>,
#> #   tumorRefCount <int>, normalAltCount <int>, normalRefCount <int>,
#> #   startPosition <int>, endPosition <int>, referenceAllele <chr>,
#> #   proteinChange <chr>, mutationType <chr>, functionalImpactScore <chr>,
#> #   fisValue <dbl>, linkXvar <chr>, linkPdb <chr>, linkMsa <chr>,
#> #   ncbiBuild <chr>, variantType <chr>, chr <chr>, variantAllele <chr>,
#> #   refseqMrnaId <chr>, proteinPosStart <int>, proteinPosEnd <int>, …
#> # ℹ Use `colnames()` to see all variable names
.get_data_by_sample(sample_study_pairs = df_pairs2, genes = 7157)
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016" and "prad_msk_2019"
#> Molecular Profile ID: blca_plasmacytoid_mskcc_2016_mutations and
#> prad_msk_2019_mutations
#> Genes: 7157
#> # A tibble: 2 × 33
#>   hugoG…¹ entre…² uniqu…³ uniqu…⁴ molec…⁵ sampl…⁶ patie…⁷ studyId center mutat…⁸
#>   <chr>     <int> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  <chr>  
#> 1 TP53       7157 UC0wMD… UC0wMD… blca_p… P-0002… P-0002… blca_p… MSK-I… NA     
#> 2 TP53       7157 UC0wMD… UC0wMD… blca_p… P-0002… P-0002… blca_p… MSK-I… NA     
#> # … with 23 more variables: validationStatus <chr>, tumorAltCount <int>,
#> #   tumorRefCount <int>, normalAltCount <int>, normalRefCount <int>,
#> #   startPosition <int>, endPosition <int>, referenceAllele <chr>,
#> #   proteinChange <chr>, mutationType <chr>, functionalImpactScore <chr>,
#> #   fisValue <dbl>, linkXvar <chr>, linkPdb <chr>, linkMsa <chr>,
#> #   ncbiBuild <chr>, variantType <chr>, keyword <chr>, chr <chr>,
#> #   variantAllele <chr>, refseqMrnaId <chr>, proteinPosStart <int>, …
#> # ℹ Use `colnames()` to see all variable names
.get_data_by_sample(sample_study_pairs = df_pairs2, data_type = "cna")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016" and "prad_msk_2019"
#> Molecular Profile ID: blca_plasmacytoid_mskcc_2016_cna and prad_msk_2019_cna
#> Genes: "All available genes"
#> NULL
.get_data_by_sample(sample_study_pairs = df_pairs2, data_type = "fusion")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016" and "prad_msk_2019"
#> Molecular Profile ID: blca_plasmacytoid_mskcc_2016_fusion and
#> prad_msk_2019_fusion
#> Genes: "All available genes"
#> # A tibble: 2 × 17
#>   uniqueSample…¹ uniqu…² molec…³ sampl…⁴ patie…⁵ studyId site1…⁶ site1…⁷ site1…⁸
#>   <chr>          <chr>   <chr>   <chr>   <chr>   <chr>     <int> <chr>   <chr>  
#> 1 UC0wMDAyMTQ2L… UC0wMD… blca_p… P-0002… P-0002… blca_p…     999 CDH1    NA     
#> 2 c19DX0NBVVdUN… cF9DX0… prad_m… s_C_CA… p_C_CA… prad_m…    7113 TMPRSS2 NA     
#> # … with 8 more variables: site1Position <int>, ncbiBuild <chr>, center <chr>,
#> #   eventInfo <chr>, variantClass <chr>, comments <chr>,
#> #   site2EntrezGeneId <int>, site2HugoSymbol <chr>, and abbreviated variable
#> #   names ¹​uniqueSampleKey, ²​uniquePatientKey, ³​molecularProfileId, ⁴​sampleId,
#> #   ⁵​patientId, ⁶​site1EntrezGeneId, ⁷​site1HugoSymbol, ⁸​site1Chromosome
#> # ℹ Use `colnames()` to see all variable names
# }