Internal Function to Get Mutations/CNA/Fusion By Sample ID

Internal Function to Get Mutations/CNA/Fusion By Sample ID

Usage

.get_data_by_sample(
  sample_id = NULL,
  study_id = NULL,
  molecular_profile_id = NULL,
  sample_study_pairs = NULL,
  data_type = c("mutation", "cna", "fusion"),
  genes = NULL,
  panel = NULL,
  add_hugo = TRUE,
  base_url = NULL
)

Arguments

sample_id

a vector of sample IDs (character)

study_id

A string indicating the study ID from which to pull data. If no study ID, will guess the study ID based on your URL and inform. Only 1 study ID can be passed. If mutations/cna from more than 1 study needed, see sample_study_pairs

molecular_profile_id

A string indicating the molecular profile ID from which to pull data. If ID supplied, will guess the molecular profile ID based on the study ID. Only 1 molecular profile ID can be passed. If mutations from more than 1 study needed, see sample_study_pairs

sample_study_pairs

A dataframe with columns: sample_id, study_id and molecular_profile_id (optional). Variations in capitalization of column names are accepted. This can be used in place of sample_id, study_id, molecular_profile_id arguments above if you need to pull samples from several different studies at once. If passed this will take overwrite sample_id, study_id, molecular_profile_id if also passed.

data_type

specify what type of data to return. Options aremutations, cna, fusion

genes

A vector of Entrez ids or Hugo symbols. If Hugo symbols are supplied, they will be converted to entrez ids using the get_entrez_id() function. If panel and genes are both NULL (default), it will return gene results for all available genomic data for that sample.

panel

One or more panel IDs to query (e.g. 'IMPACT468'). If panel and genes are both NULL (default), it will return gene results for all available genomic data for that sample.

add_hugo

Logical indicating whether HugoSymbol should be added to your results. cBioPortal API does not return this by default (only EntrezId) but this function's default is TRUE and adds this by default.

base_url

The database URL to query If NULL will default to URL set with set_cbioportal_db(<your_db>)

Value

A dataframe of mutations or CNAs

Examples

set_cbioportal_db("public")
#> ✔ You are successfully connected!
#> ✔ base_url for this R session is now set to "www.cbioportal.org/api" 
.get_data_by_sample(sample_id = c("TCGA-OR-A5J2-01","TCGA-OR-A5J6-01"),
 study_id = "acc_tcga", data_type = "mutation")
#> The following parameters were used in query:
#> Study ID: "acc_tcga"
#> Molecular Profile ID: "acc_tcga_mutations"
#> Genes: "All available genes"
#> # A tibble: 173 × 33
#>    hugoGeneSymbol entrezGeneId uniqueSampleKey uniquePatientKey molecularProfil…
#>    <chr>                 <int> <chr>           <chr>            <chr>           
#>  1 ZFPM1                161882 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  2 ZFPM1                161882 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  3 ZFPM1                161882 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  4 ZFPM1                161882 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  5 ZNF787               126208 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  6 PODXL                  5420 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  7 CCDC102A              92922 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  8 TVP23C               201158 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#>  9 RINL                 126432 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#> 10 ZNF628                89887 VENHQS1PUi1BNU… VENHQS1PUi1BNUo… acc_tcga_mutati…
#> # … with 163 more rows, and 28 more variables: sampleId <chr>, patientId <chr>,
#> #   studyId <chr>, center <chr>, mutationStatus <chr>, validationStatus <chr>,
#> #   tumorAltCount <int>, tumorRefCount <int>, normalAltCount <int>,
#> #   normalRefCount <int>, startPosition <int>, endPosition <int>,
#> #   referenceAllele <chr>, proteinChange <chr>, mutationType <chr>,
#> #   functionalImpactScore <chr>, fisValue <dbl>, linkXvar <chr>, linkPdb <chr>,
#> #   linkMsa <chr>, ncbiBuild <chr>, variantType <chr>, keyword <chr>, …

.get_data_by_sample(sample_id = c("DS-sig-010-P2"),
 molecular_profile_id = "blca_plasmacytoid_mskcc_2016_cna", data_type = "cna")
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016"
#> Molecular Profile ID: "blca_plasmacytoid_mskcc_2016_cna"
#> Genes: "All available genes"
#> # A tibble: 2 × 9
#>   hugoGeneSymbol entrezGeneId uniqueSampleKey  uniquePatientKey molecularProfil…
#>   <chr>                 <int> <chr>            <chr>            <chr>           
#> 1 YAP1                  10413 RFMtc2lnLTAxMC1… RFMtc2lnLTAxMDp… blca_plasmacyto…
#> 2 CD79B                   974 RFMtc2lnLTAxMC1… RFMtc2lnLTAxMDp… blca_plasmacyto…
#> # … with 4 more variables: sampleId <chr>, patientId <chr>, studyId <chr>,
#> #   alteration <int>

.get_data_by_sample(sample_id = c("P-0002146-T01-IM3"),
 study_id = "blca_plasmacytoid_mskcc_2016", data_type = "fusion")
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016"
#> Molecular Profile ID: "blca_plasmacytoid_mskcc_2016_fusion"
#> Genes: "All available genes"
#> # A tibble: 1 × 15
#>   uniqueSampleKey   uniquePatientKey molecularProfil… sampleId patientId studyId
#>   <chr>             <chr>            <chr>            <chr>    <chr>     <chr>  
#> 1 UC0wMDAyMTQ2LVQw… UC0wMDAyMTQ2OmJ… blca_plasmacyto… P-00021… P-0002146 blca_p…
#> # … with 9 more variables: site1EntrezGeneId <int>, site1HugoSymbol <chr>,
#> #   site1Chromosome <chr>, site1Position <int>, ncbiBuild <chr>, center <chr>,
#> #   eventInfo <chr>, variantClass <chr>, comments <chr>

# \donttest{
df_pairs <- data.frame(
"sample_id" = c("s_C_36924L_P001_d",
"s_C_03LNU8_P001_d"),
"study_id" = c("prad_msk_2019"))

.get_data_by_sample(sample_study_pairs = df_pairs, data_type = "mutation")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "prad_msk_2019"
#> Molecular Profile ID: prad_msk_2019_mutations
#> Genes: "All available genes"
#> # A tibble: 1 × 33
#>   hugoGeneSymbol entrezGeneId uniqueSampleKey  uniquePatientKey molecularProfil…
#>   <chr>                 <int> <chr>            <chr>            <chr>           
#> 1 TP53                   7157 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… prad_msk_2019_m…
#> # … with 28 more variables: sampleId <chr>, patientId <chr>, studyId <chr>,
#> #   center <chr>, mutationStatus <chr>, validationStatus <chr>,
#> #   tumorAltCount <int>, tumorRefCount <int>, normalAltCount <int>,
#> #   normalRefCount <int>, startPosition <int>, endPosition <int>,
#> #   referenceAllele <chr>, proteinChange <chr>, mutationType <chr>,
#> #   functionalImpactScore <chr>, fisValue <dbl>, linkXvar <chr>, linkPdb <chr>,
#> #   linkMsa <chr>, ncbiBuild <chr>, variantType <chr>, keyword <chr>, …
.get_data_by_sample(sample_study_pairs = df_pairs, genes = 7157, data_type = "mutation")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "prad_msk_2019"
#> Molecular Profile ID: prad_msk_2019_mutations
#> Genes: 7157
#> # A tibble: 1 × 33
#>   hugoGeneSymbol entrezGeneId uniqueSampleKey  uniquePatientKey molecularProfil…
#>   <chr>                 <int> <chr>            <chr>            <chr>           
#> 1 TP53                   7157 c19DXzAzTE5VOF9… cF9DXzAzTE5VODp… prad_msk_2019_m…
#> # … with 28 more variables: sampleId <chr>, patientId <chr>, studyId <chr>,
#> #   center <chr>, mutationStatus <chr>, validationStatus <chr>,
#> #   tumorAltCount <int>, tumorRefCount <int>, normalAltCount <int>,
#> #   normalRefCount <int>, startPosition <int>, endPosition <int>,
#> #   referenceAllele <chr>, proteinChange <chr>, mutationType <chr>,
#> #   functionalImpactScore <chr>, fisValue <dbl>, linkXvar <chr>, linkPdb <chr>,
#> #   linkMsa <chr>, ncbiBuild <chr>, variantType <chr>, keyword <chr>, …
.get_data_by_sample(sample_study_pairs = df_pairs, data_type = "cna")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "prad_msk_2019"
#> Molecular Profile ID: prad_msk_2019_cna
#> Genes: "All available genes"
#> # A tibble: 1 × 9
#>   hugoGeneSymbol entrezGeneId uniqueSampleKey  uniquePatientKey molecularProfil…
#>   <chr>                 <int> <chr>            <chr>            <chr>           
#> 1 PTEN                   5728 c19DXzM2OTI0TF9… cF9DXzM2OTI0TDp… prad_msk_2019_c…
#> # … with 4 more variables: sampleId <chr>, patientId <chr>, studyId <chr>,
#> #   alteration <int>
.get_data_by_sample(sample_study_pairs = df_pairs, data_type = "fusion")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "prad_msk_2019"
#> Molecular Profile ID: prad_msk_2019_fusion
#> Genes: "All available genes"
#> # A tibble: 0 × 0

df_pairs2 <- data.frame(
"sample_id" = c("P-0002146-T01-IM3", "s_C_CAUWT7_P001_d"),
 "study_id" = c("blca_plasmacytoid_mskcc_2016", "prad_msk_2019"))

.get_data_by_sample(sample_study_pairs = df_pairs2, data_type = "mutation")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016" and "prad_msk_2019"
#> Molecular Profile ID: blca_plasmacytoid_mskcc_2016_mutations and
#> prad_msk_2019_mutations
#> Genes: "All available genes"
#> # A tibble: 7 × 33
#>   hugoGeneSymbol entrezGeneId uniqueSampleKey  uniquePatientKey molecularProfil…
#>   <chr>                 <int> <chr>            <chr>            <chr>           
#> 1 TERT                   7015 UC0wMDAyMTQ2LVQ… UC0wMDAyMTQ2OmJ… blca_plasmacyto…
#> 2 TP53                   7157 UC0wMDAyMTQ2LVQ… UC0wMDAyMTQ2OmJ… blca_plasmacyto…
#> 3 NOTCH4                 4855 UC0wMDAyMTQ2LVQ… UC0wMDAyMTQ2OmJ… blca_plasmacyto…
#> 4 BLM                     641 UC0wMDAyMTQ2LVQ… UC0wMDAyMTQ2OmJ… blca_plasmacyto…
#> 5 TP53                   7157 UC0wMDAyMTQ2LVQ… UC0wMDAyMTQ2OmJ… blca_plasmacyto…
#> 6 CDKN1A                 1026 UC0wMDAyMTQ2LVQ… UC0wMDAyMTQ2OmJ… blca_plasmacyto…
#> 7 RB1                    5925 UC0wMDAyMTQ2LVQ… UC0wMDAyMTQ2OmJ… blca_plasmacyto…
#> # … with 28 more variables: sampleId <chr>, patientId <chr>, studyId <chr>,
#> #   center <chr>, mutationStatus <chr>, validationStatus <chr>,
#> #   tumorAltCount <int>, tumorRefCount <int>, normalAltCount <int>,
#> #   normalRefCount <int>, startPosition <int>, endPosition <int>,
#> #   referenceAllele <chr>, proteinChange <chr>, mutationType <chr>,
#> #   functionalImpactScore <chr>, fisValue <dbl>, linkXvar <chr>, linkPdb <chr>,
#> #   linkMsa <chr>, ncbiBuild <chr>, variantType <chr>, chr <chr>, …
.get_data_by_sample(sample_study_pairs = df_pairs2, genes = 7157)
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016" and "prad_msk_2019"
#> Molecular Profile ID: blca_plasmacytoid_mskcc_2016_mutations and
#> prad_msk_2019_mutations
#> Genes: 7157
#> # A tibble: 2 × 33
#>   hugoGeneSymbol entrezGeneId uniqueSampleKey  uniquePatientKey molecularProfil…
#>   <chr>                 <int> <chr>            <chr>            <chr>           
#> 1 TP53                   7157 UC0wMDAyMTQ2LVQ… UC0wMDAyMTQ2OmJ… blca_plasmacyto…
#> 2 TP53                   7157 UC0wMDAyMTQ2LVQ… UC0wMDAyMTQ2OmJ… blca_plasmacyto…
#> # … with 28 more variables: sampleId <chr>, patientId <chr>, studyId <chr>,
#> #   center <chr>, mutationStatus <chr>, validationStatus <chr>,
#> #   tumorAltCount <int>, tumorRefCount <int>, normalAltCount <int>,
#> #   normalRefCount <int>, startPosition <int>, endPosition <int>,
#> #   referenceAllele <chr>, proteinChange <chr>, mutationType <chr>,
#> #   functionalImpactScore <chr>, fisValue <dbl>, linkXvar <chr>, linkPdb <chr>,
#> #   linkMsa <chr>, ncbiBuild <chr>, variantType <chr>, keyword <chr>, …
.get_data_by_sample(sample_study_pairs = df_pairs2, data_type = "cna")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016" and "prad_msk_2019"
#> Molecular Profile ID: blca_plasmacytoid_mskcc_2016_cna and prad_msk_2019_cna
#> Genes: "All available genes"
#> NULL
.get_data_by_sample(sample_study_pairs = df_pairs2, data_type = "fusion")
#> Joining, by = "study_id"
#> The following parameters were used in query:
#> Study ID: "blca_plasmacytoid_mskcc_2016" and "prad_msk_2019"
#> Molecular Profile ID: blca_plasmacytoid_mskcc_2016_fusion and
#> prad_msk_2019_fusion
#> Genes: "All available genes"
#> # A tibble: 2 × 17
#>   uniqueSampleKey   uniquePatientKey molecularProfil… sampleId patientId studyId
#>   <chr>             <chr>            <chr>            <chr>    <chr>     <chr>  
#> 1 UC0wMDAyMTQ2LVQw… UC0wMDAyMTQ2OmJ… blca_plasmacyto… P-00021… P-0002146 blca_p…
#> 2 c19DX0NBVVdUN19Q… cF9DX0NBVVdUNzp… prad_msk_2019_f… s_C_CAU… p_C_CAUW… prad_m…
#> # … with 11 more variables: site1EntrezGeneId <int>, site1HugoSymbol <chr>,
#> #   site1Chromosome <chr>, site1Position <int>, ncbiBuild <chr>, center <chr>,
#> #   eventInfo <chr>, variantClass <chr>, comments <chr>,
#> #   site2EntrezGeneId <int>, site2HugoSymbol <chr>
# }