Get copy number alterations by sample ID or study ID

Get copy number alterations by sample ID or study ID

get_cna(sample_id = NULL, study_id = NULL, panel = NULL, genes = NULL)

Arguments

sample_id

A character vector of sample ids. Can be NULL if a study_id is passed. If sample IDs are passed with no study_id, IMPACT studies will be queried by default. If non IMPACT ID samples passed, a study_id is required as well to retrieve data.

study_id

A character vector study IDs. See get_studies() to see available studies. If passed with no specified sample_id all samples for specified studies will be returned.

panel

OPTIONAL argument. A character vector of length 1 indicating a specific panel to be used. If not NULL, the panel will be looked up with get_panel() and only genes in that panel will be returned.

genes

A list of genes to query. default is all impact genes.

Value

A dataframe of CNA for each sample ID (in maf file format)

Examples

get_cbioportal_db('public')
get_cna(sample_id = c("P-0000004-T01-IM3", "P-0000015-T01-IM3"))
#> No genes specified by user, so querying curated list of impact genes.
#>               To view queried gene list run `cbioportalR::impact_gene_info`
#> Warning: If you are an MSK researcher, for most up to date IMPACT data you should connect to MSK cbioportal. 
#> This function is using limited public IMPACT data (study_id = 'msk_impact_2017')
#> Warning: No CBIOPORTAL_TOKEN in `.Renviron`. Try `usethis::edit_r_environ()` to add a token
#> # A tibble: 14 × 9
#>    uniqueSampleKey          uniquePatientKey molecularProfil… sampleId patientId
#>    <chr>                    <chr>            <chr>            <chr>    <chr>    
#>  1 UC0wMDAwMDA0LVQwMS1JTTM… UC0wMDAwMDA0Om1… msk_impact_2017… P-00000… P-0000004
#>  2 UC0wMDAwMDA0LVQwMS1JTTM… UC0wMDAwMDA0Om1… msk_impact_2017… P-00000… P-0000004
#>  3 UC0wMDAwMDA0LVQwMS1JTTM… UC0wMDAwMDA0Om1… msk_impact_2017… P-00000… P-0000004
#>  4 UC0wMDAwMDE1LVQwMS1JTTM… UC0wMDAwMDE1Om1… msk_impact_2017… P-00000… P-0000015
#>  5 UC0wMDAwMDE1LVQwMS1JTTM… UC0wMDAwMDE1Om1… msk_impact_2017… P-00000… P-0000015
#>  6 UC0wMDAwMDE1LVQwMS1JTTM… UC0wMDAwMDE1Om1… msk_impact_2017… P-00000… P-0000015
#>  7 UC0wMDAwMDE1LVQwMS1JTTM… UC0wMDAwMDE1Om1… msk_impact_2017… P-00000… P-0000015
#>  8 UC0wMDAwMDE1LVQwMS1JTTM… UC0wMDAwMDE1Om1… msk_impact_2017… P-00000… P-0000015
#>  9 UC0wMDAwMDE1LVQwMS1JTTM… UC0wMDAwMDE1Om1… msk_impact_2017… P-00000… P-0000015
#> 10 UC0wMDAwMDE1LVQwMS1JTTM… UC0wMDAwMDE1Om1… msk_impact_2017… P-00000… P-0000015
#> 11 UC0wMDAwMDE1LVQwMS1JTTM… UC0wMDAwMDE1Om1… msk_impact_2017… P-00000… P-0000015
#> 12 UC0wMDAwMDE1LVQwMS1JTTM… UC0wMDAwMDE1Om1… msk_impact_2017… P-00000… P-0000015
#> 13 UC0wMDAwMDE1LVQwMS1JTTM… UC0wMDAwMDE1Om1… msk_impact_2017… P-00000… P-0000015
#> 14 UC0wMDAwMDE1LVQwMS1JTTM… UC0wMDAwMDE1Om1… msk_impact_2017… P-00000… P-0000015
#> # … with 4 more variables: entrezGeneId <int>, studyId <chr>, alteration <int>,
#> #   Hugo_Symbol <chr>