Get all genetic data for a study or list of IDs

Get all genetic data for a study or list of IDs

get_genetics(
  sample_ids = NULL,
  study_id = NULL,
  genes = NULL,
  panel = NULL,
  mutations = TRUE,
  fusions = TRUE,
  cna = TRUE,
  seg = FALSE
)

Arguments

sample_ids
A character vector of sample ids

study_id
A character vector of length 1 specifying study to query. See get_studies() to see list of available studies.

genes
A list of genes to query. default is all impact genes.

panel
OPTIONAL argument. A character vector of length 1 indicating a specific panel to be used. If not NULL, the panel will be looked up with get_panel() and only genes in that panel will be returned.

mutations
Boolean specifying if mutation data should be fetched. Default is TRUE.

fusions
Boolean specifying if fusion data should be fetched. Default is TRUE.

cna
Boolean specifying if cna data should be fetched. Default is TRUE.

seg
Boolean specifying if segmentation data should be fectched. Default is FALSE.

Value

A list of dataframes containing genetic data.

Examples

if (FALSE) {
# IMPACT #
get_genetics(sample_ids = c("P-0000004-T01-IM3", "P-0000012-T02-IM3"),
genes = 207)

# TCGA #
get_genetics(sample_ids =  c("TCGA-17-Z023-01","TCGA-02-0003-01","TCGA-02-0055-01"),
 database = "tcga")
}