Get all genetic data for a study or list of IDs

Get all genetic data for a study or list of IDs

get_genetics(
  sample_ids = NULL,
  study_id = NULL,
  genes = NULL,
  panel = NULL,
  mutations = TRUE,
  fusions = TRUE,
  cna = TRUE
)

Arguments

sample_ids
A character vector of sample ids. Can be NULL if a study_id is passed. If sample IDs are passed with no study_id, IMPACT studies will be queried by default. If non IMPACT ID samples passed, a study_id is required as well to retrieve data.

study_id
A character vector study IDs. See get_studies() to see available studies. If passed with no specified sample_id all samples for specified studies will be returned.

genes
A vector of Entrez IDs of genes to query. default is all impact genes.

panel
OPTIONAL argument. A character vector of length 1 indicating a specific panel to be used. If not NULL, the panel will be looked up with get_panel() and only genes in that panel will be returned.

mutations
Boolean specifying if mutation data should be fetched. Default is TRUE.

fusions
Boolean specifying if fusion data should be fetched. Default is TRUE.

cna
Boolean specifying if cna data should be fetched. Default is TRUE.

Value

A list of dataframes containing genetic data.

Examples

get_cbioportal_db("public")
get_genetics(sample_id = c("P-0005217-T03-IM5", "P-0038798-T01-IM6"))

#> No genes specified by user, so querying curated list of impact genes.
#>               To view queried gene list run `cbioportalR::impact_gene_info`

#> Warning: If you are an MSK researcher, for most up to date IMPACT data you should connect to MSK cbioportal. 
#> This function is using limited public IMPACT data (study_id = 'msk_impact_2017')

#> No genes specified by user, so querying curated list of impact genes.
#>               To view queried gene list run `cbioportalR::impact_gene_info`

#> Warning: If you are an MSK researcher, for most up to date IMPACT data you should connect to MSK cbioportal. 
#> This function is using limited public IMPACT data (study_id = 'msk_impact_2017')

#> $mut
#> # A tibble: 2 x 33
#>   uniqueSampleKey  uniquePatientKey  molecularProfil… Tumor_Sample_Ba… patientId
#>   <chr>            <chr>             <chr>            <chr>            <chr>    
#> 1 UC0wMDA1MjE3LVQ… UC0wMDA1MjE3Om1z… msk_impact_2017… P-0005217-T03-I… P-0005217
#> 2 UC0wMDA1MjE3LVQ… UC0wMDA1MjE3Om1z… msk_impact_2017… P-0005217-T03-I… P-0005217
#> # … with 28 more variables: entrezGeneId <int>, studyId <chr>, center <chr>,
#> #   Mutation_Status <chr>, validationStatus <chr>, tumorAltCount <int>,
#> #   tumorRefCount <int>, startPosition <int>, endPosition <int>,
#> #   referenceAllele <chr>, proteinChange <chr>, Variant_Classification <chr>,
#> #   functionalImpactScore <chr>, fisValue <dbl>, linkXvar <chr>, linkPdb <chr>,
#> #   linkMsa <chr>, ncbiBuild <chr>, Variant_Type <chr>, chr <chr>,
#> #   variantAllele <chr>, refseqMrnaId <chr>, proteinPosStart <int>,
#> #   proteinPosEnd <int>, keyword <chr>, HGVSp_Short <chr>,
#> #   Protein_position <int>, Hugo_Symbol <chr>
#> 
#> $cna
#> # A tibble: 0 x 0
#>