Get gene mutations by sample ID or Study

Get gene mutations by sample ID or Study

Usage

get_mutations(sample_id = NULL, study_id = NULL, panel = NULL, genes = NULL)

Arguments

sample_id

A character vector of sample ids. Can be NULL if a study_id is passed. If sample IDs are passed with no study_id, IMPACT studies will be queried by default. If non IMPACT ID samples passed, a study_id is required as well to retrieve data.

study_id

A character vector study IDs. See get_studies() to see available studies. If passed with no specified sample_id all samples for specified studies will be returned.

panel

OPTIONAL argument. A character vector of length 1 indicating a specific panel to be used. If not NULL, the panel will be looked up with get_panel() and only genes in that panel will be returned.

genes

A list of genes to query. default is all impact genes.

Value

A dataframe of mutations for each sample ID (in maf file format)

Examples

get_cbioportal_db("public")
get_mutations(sample_id = c("P-0005217-T03-IM5", "P-0038798-T01-IM6"))
#> No genes specified by user, so querying curated list of impact genes.
#>               To view queried gene list run `cbioportalR::impact_gene_info`
#> Warning: If you are an MSK researcher, for most up to date IMPACT data you should connect to MSK cbioportal. 
#> This function is using limited public IMPACT data (study_id = 'msk_impact_2017')
#> Warning: No CBIOPORTAL_TOKEN in `.Renviron`. Try `usethis::edit_r_environ()` to add a token
#> # A tibble: 2 × 33
#>   uniqueSampleKey           uniquePatientKey molecularProfil… sampleId patientId
#>   <chr>                     <chr>            <chr>            <chr>    <chr>    
#> 1 UC0wMDA1MjE3LVQwMy1JTTU6… UC0wMDA1MjE3Om1… msk_impact_2017… P-00052… P-0005217
#> 2 UC0wMDA1MjE3LVQwMy1JTTU6… UC0wMDA1MjE3Om1… msk_impact_2017… P-00052… P-0005217
#> # … with 28 more variables: entrezGeneId <int>, studyId <chr>, center <chr>,
#> #   mutationStatus <chr>, validationStatus <chr>, tumorAltCount <int>,
#> #   tumorRefCount <int>, startPosition <int>, endPosition <int>,
#> #   referenceAllele <chr>, proteinChange <chr>, mutationType <chr>,
#> #   functionalImpactScore <chr>, fisValue <dbl>, linkXvar <chr>, linkPdb <chr>,
#> #   linkMsa <chr>, ncbiBuild <chr>, variantType <chr>, keyword <chr>,
#> #   chr <chr>, variantAllele <chr>, refseqMrnaId <chr>, …